UI: the shared text cut sizes its first guess on a float ratio, not an int one
Every elided string in the game came out as a single letter followed by an
ellipsis - a ship name in the colonies list, a status cell, a patrol's worlds
(bench 621).

The first guess at how many characters fit divides the available room by the
text's width. Both are ints, so the division is an integer one and yields 0 for
any text wider than its room: the guess becomes one character, which then fits,
so the loop that would refine it never runs and the back-off to a whole word has
nothing to work with.

The ratio is now a float. The four callers that elide - the queued item's name,
the ship list's status, a patrol's colony list, and the fold of a table column -
all get a real cut back.

+++ Ship_Game/UI/UITable.cs
@@ -155,7 +155,11 @@ namespace Ship_Game.UI
         {
             if (text.IsEmpty() || font.TextWidth(text) <= room)
                 return text;
-            int len = Math.Max(1, (int)(text.Length * (room / font.TextWidth(text))));
+            // ⚠ the ratio is a FLOAT one: room and TextWidth are both int, so an integer division
+            // here yields 0 for every text wider than the room - the first guess becomes one
+            // character, "F...", which then fits and stops the loop before it can do better. Every
+            // cut in the game came out as a single letter (bench 621).
+            int len = Math.Max(1, (int)(text.Length * ((float)room / font.TextWidth(text))));
             string t = text.Substring(0, len).TrimEnd() + "...";
             while (len > 1 && font.TextWidth(t) > room)
                 t = text.Substring(0, --len).TrimEnd() + "...";
